requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
easyocr==1.7.0
rapidfuzz==3.3.1
Pillow==10.0.1
numpy==1.24.3
